+++ packages/rpc/src/lib/transactions/index.ts
@@ -1,7 +1,7 @@
 import type { ApiPromise } from '@polkadot/api'
 import type { SubmittableExtrinsic } from '@polkadot/api/types'
 import type { ISubmittableResult } from '@polkadot/types/types'
-import * as Te from '@zeitgeistpm/utility/dist/esm/taskeither'
+import * as Te from '@zeitgeistpm/utility/dist/cjs/taskeither'
 import { isExtSigner, KeyringPairOrExtSigner } from '../keyring'
 import { RetractedError, TransactionError, UnknownDispatchError } from './types'