+++ packages/sdk/src/meta/market/codec.ts
@@ -1,6 +1,6 @@
 import type { AnyJson } from '@polkadot/types/types'
-import type { Codec } from '@zeitgeistpm/utility/dist/esm/codec'
-import { either, left, right } from '@zeitgeistpm/utility/dist/esm/either'
+import type { Codec } from '@zeitgeistpm/utility/dist/cjs/codec'
+import { either, left, right } from '@zeitgeistpm/utility/dist/cjs/either'
 import { StructError, validate } from 'superstruct'
 import { IOMarketMetadata, MarketMetadata } from '.'