+++ packages/sdk/src/model/markets/types.ts
@@ -1,5 +1,5 @@
-import { PFunc } from '@zeitgeistpm/utility/dist/esm/pfunc'
-import * as Te from '@zeitgeistpm/utility/dist/esm/taskeither'
+import { PFunc } from '@zeitgeistpm/utility/dist/cjs/pfunc'
+import * as Te from '@zeitgeistpm/utility/dist/cjs/taskeither'
 import { Observable } from 'rxjs'
 import { Context, RpcContext } from '../../context'
 import { MetadataStorage } from '../../meta'