+++ packages/augment-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zeitgeistpm/augment-api
 
+## 2.0.24
+
+### Patch Changes
+
+- asd
+
 ## 2.0.23
 
 ### Patch Changes